ai_agents/requirements.txt
gitpython
faiss-cpu
transformers
torch
sentence-transformers
PyGithub
pandas
jinja2
weasyprint
ollama